[#377] Removed added pinned simplequantity copy from HuskyFhirValidatorImpl and moved interceptor handling from ChEmedEprHapiValidator to HuskyFhirValidatorImpl, from which it now extends.
