+++ husky-fhir/husky-fhir-emed-ch/husky-fhir-emed-ch-epr/src/main/java/org/projecthusky/fhir/emed/ch/epr/validator/ChEmedEprHapiValidator.java
@@ -1,56 +1,27 @@
 package org.projecthusky.fhir.emed.ch.epr.validator;
 
 import ca.uhn.fhir.context.FhirContext;
-import ca.uhn.fhir.context.support.DefaultProfileValidationSupport;
-import ca.uhn.fhir.rest.server.interceptor.validation.ValidationMessagePostProcessingInterceptor;
-import ca.uhn.fhir.validation.FhirValidator;
 import ca.uhn.fhir.validation.ValidationOptions;
-import lombok.Setter;
 import lombok.extern.slf4j.Slf4j;
-import org.checkerframework.checker.nullness.qual.NonNull;
 import org.checkerframework.checker.nullness.qual.Nullable;
 import org.hl7.fhir.common.hapi.validation.support.*;
-import org.hl7.fhir.common.hapi.validation.validator.FhirInstanceValidator;
 import org.hl7.fhir.r4.model.*;
 import org.projecthusky.fhir.emed.ch.epr.resource.ChEmedEprDocument;
 import org.projecthusky.fhir.emed.ch.epr.validator.logicvalidator.LogicValidator;
 import org.projecthusky.fhir.validation.HuskyFhirValidator;
-import org.projecthusky.fhir.validation.logging.ValidationResultLogger;
+import org.projecthusky.fhir.validation.impl.HuskyFhirValidatorImpl;
 import org.projecthusky.fhir.validation.model.ValidationResult;
 
 import java.io.IOException;
 import java.io.InputStream;
 import java.nio.charset.StandardCharsets;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Objects;
 
 /**
  * A HAPI based implementation of the {@link ChEmedEprValidator}.
  */
 @Slf4j
-public class ChEmedEprHapiValidator implements ChEmedEprValidator{
-    final private FhirValidator validator;
+public class ChEmedEprHapiValidator extends HuskyFhirValidatorImpl implements ChEmedEprValidator {
     final private LogicValidator logicValidator = new LogicValidator();
-
-    /**
-     * Whether the validation outcome should be logged with logger when calling a validation method.
-     */
-    @Setter
-    private boolean logValidationOutcome = true;
-
-    /**
-     * The logger to use in case validation result logger is enabled.
-     */
-    @Setter
-    private @Nullable ValidationResultLogger validationResultLogger;
-
-    /**
-     * A list of validation message post-processing interceptors. These interceptors are handled right after performing
-     * the base HAPI validation but before (optionally) logging and then returning the result.
-     */
-    @Setter
-    private List<@NonNull ValidationMessagePostProcessingInterceptor> interceptors = new ArrayList<>();
 
     /**
      * Creates an instance of an <i>offline</i> {@link ChEmedEprValidator}, that is, a validator that will not try to
@@ -71,39 +42,11 @@ public class ChEmedEprHapiValidator implements ChEmedEprValidator{
      * @throws IOException if the NPM packages can't be found in the classpath.
      */
     public ChEmedEprHapiValidator(final FhirContext context, final @Nullable String txServer)  throws IOException {
-        final var packageSupport = new NpmPackageValidationSupport(context);
-        for (var packageResource : packageResourceList) packageSupport.loadPackageFromClasspath(packageResource);
-        final var prepopulatedValidationSupport = new PrePopulatedValidationSupport(context);
-        final var chain = new ValidationSupportChain(
-                prepopulatedValidationSupport,
-                packageSupport,
-                new DefaultProfileValidationSupport(context),
-                new SnapshotGeneratingValidationSupport(context),
-                (txServer != null && !txServer.isBlank())? new RemoteTerminologyServiceValidationSupport(context, txServer)
-                        : new InMemoryTerminologyServerValidationSupport(context),
-                new CommonCodeSystemsTerminologyService(context)
-        );
+        super(context, packageResourceList, txServer);
+
         final var def = chain.fetchStructureDefinition("http://hl7.org/fhir/StructureDefinition/SimpleQuantity");
-        prepopulatedValidationSupport.addStructureDefinition(((StructureDefinition) def).copy().setUrl("http://hl7.org/fhir/StructureDefinition/SimpleQuantity|4.0.1"));
-        validator = context.newValidator();
-        final var instanceValidator = new FhirInstanceValidator(chain);
-        validator.registerValidatorModule(instanceValidator);
-    }
-
-    /**
-     * Gets the list of validation message post-processing interceptors to be processed after validation.
-     */
-    public List<@NonNull ValidationMessagePostProcessingInterceptor>  getInterceptors() {
-        if (interceptors == null) interceptors = new ArrayList<>();
-        return interceptors;
-    }
-
-    /**
-     * Adds an interceptor to the list of validation message post-processing interceptors.
-     */
-    public ChEmedEprHapiValidator addValidationMessagePostProcessingInterceptor(final ValidationMessagePostProcessingInterceptor interceptor) {
-        getInterceptors().add(interceptor);
-        return this;
+        prepopulatedValidationSupport
+                .addStructureDefinition(((StructureDefinition) def).copy().setUrl("http://hl7.org/fhir/StructureDefinition/SimpleQuantity|4.0.1"));
     }
 
     @Override
@@ -120,46 +63,5 @@ public class ChEmedEprHapiValidator implements ChEmedEprValidator{
             huskyResult.add(logicValidator.validate(document));
         logValidationResult(huskyResult);
         return huskyResult;
-    }
-
-    @Override
-    public ValidationResult validateDocumentBundle(final Bundle bundle, final String profile) {
-        final var validationOptions = new ValidationOptions();
-        validationOptions.addProfile(Objects.requireNonNull(profile));
-        final var result = validator.validateWithResult(Objects.requireNonNull(bundle), validationOptions);
-        handleValidationResult(result);
-        final var huskyResult =
-        		HuskyFhirValidator.toHuskyValidationResult((OperationOutcome) result.toOperationOutcome());
-        logValidationResult(huskyResult);
-        return huskyResult;
-    }
-
-    /**
-     * Handles the HAPI FHIR validation result by having all registered post-processors handle it. The result is
-     * expected to be potentially modified by the handling. No new instance will be kept, results must be modified.
-     */
-    protected void handleValidationResult(final ca.uhn.fhir.validation.ValidationResult result) {
-        for (final var interceptor : getInterceptors()) interceptor.handle(result);
-    }
-
-    /**
-     * Logs the validation result if enabled in the validator config. If enabled and no explicit
-     * {@link ValidationResultLogger} has been set, a default one will be automatically set and used.
-     *
-     * @param validationResult The validation result to be logged.
-     */
-    protected void logValidationResult(final ValidationResult validationResult) {
-        if (logValidationOutcome) {
-            if (validationResultLogger == null) validationResultLogger = getDefaultValidationResultLogger();
-            if (validationResultLogger != null) validationResultLogger.logValidationResult(validationResult);
-            else log.error("The CH EMED EPR validator is set to log results, but no logger is set. Cannot log results.");
-        }
-    }
-
-    /**
-     * Gets a default validation result logger.
-     */
-    public ValidationResultLogger getDefaultValidationResultLogger() {
-        return new ValidationResultLogger(log);
     }
 }
